requirements.txt
langchain==0.0.348
python-dotenv==1.0.0
streamlit==1.29.0